Update anchor link to recipe on chicken wings page
Opening the recipe link would send user to the recipe on the same tab,
making it a bit inconvenient when user wants to go back to the page.
Added 'target=_blank' so that opening the link would open a new tab to
the page.

Additionally added 'rel=noopener noreferrer' to prevent security issues
from using the 'target' attribute.